requirements.txt
# CivicCam Requirements
# AI-Powered Littering Detection System

# Core ML/AI
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.6.0
pillow>=9.0.0

# OCR
easyocr>=1.7.0

# Web Interface
streamlit>=1.28.0

# Telegram Bot
python-telegram-bot>=20.0

# Data Processing
numpy>=1.23.0
pandas>=1.5.0
scipy>=1.9.0

# Visualization
matplotlib>=3.6.0
plotly>=5.0.0

# Database
# sqlite3 is built-in

# Utilities
pyyaml>=6.0
requests>=2.28.0
tqdm>=4.64.0

# Dataset Management
roboflow>=1.0.0
